# contracts/activation-contract/src/prism_activation_contract/openapi.yaml
openapi: 3.1.0
info:
  title: PRISM Activation Contract
  version: 1.0.0
  description: |
    Warehouse-agnostic activation surface — PRISM's namesake.

    One gold table, many warehouses: activate gold table X into warehouse Y,
    then query table X regardless of which warehouse currently serves it.

    Adapters (Redshift Serverless zero-ETL/COPY, Snowflake Horizon Catalog /
    Iceberg REST zero-copy) implement this contract behind `activation-gateway/`.
    CI uses local mocked warehouse endpoints (ADR-001).
servers:
  - url: http://localhost:9103
    description: Local activation-gateway
tags:
  - name: health
  - name: activate
  - name: query
  - name: routing
paths:
  /health:
    get:
      tags: [health]
      summary: Gateway health
      operationId: health
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /v1/warehouses:
    get:
      tags: [routing]
      summary: List supported warehouse adapters
      operationId: listWarehouses
      responses:
        "200":
          description: Warehouse catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehouseListResponse"

  /v1/activate:
    post:
      tags: [activate]
      summary: Activate gold table X into warehouse Y
      operationId: activateGoldTable
      description: |
        Materialize or register a lakehouse gold table into the target warehouse.
        Redshift prefers zero-ETL/auto-copy with Parquet/Iceberg COPY as fallback.
        Snowflake registers an Iceberg REST / Horizon Catalog zero-copy read against
        the same S3 (or local) gold URI — storage is not duplicated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateRequest"
            examples:
              redshift:
                value:
                  gold_table: asset_daily_metrics
                  warehouse: redshift
                  gold_uri: file:///data/lakehouse/gold/asset_daily_metrics
                  strategy: auto
              snowflake:
                value:
                  gold_table: asset_daily_metrics
                  warehouse: snowflake
                  gold_uri: file:///data/lakehouse/gold/asset_daily_metrics
                  strategy: iceberg_rest
      responses:
        "200":
          description: Activation succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivateResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Gold URI or table not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/query:
    post:
      tags: [query]
      summary: Query table X regardless of serving warehouse
      operationId: queryActivatedTable
      description: |
        Execute a bounded read against an activated gold table. When `warehouse`
        is `auto` (default), the gateway routes to the warehouse currently marked
        as primary for that table. Callers may pin `redshift` or `snowflake`
        for conformance / failover checks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
            example:
              table: asset_daily_metrics
              warehouse: auto
              sql: >
                SELECT asset_id, metric_date, ping_count
                FROM asset_daily_metrics
                ORDER BY asset_id
              limit: 100
      responses:
        "200":
          description: Query result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "400":
          description: Invalid SQL or request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Table not activated on the requested warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/routing/{table}:
    get:
      tags: [routing]
      summary: Which warehouse(s) currently serve table X
      operationId: getTableRouting
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-z][a-z0-9_]{1,63}$'
      responses:
        "200":
          description: Routing map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingResponse"
        "404":
          description: Table has no activations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    WarehouseId:
      type: string
      enum: [redshift, snowflake]

    WarehouseTarget:
      type: string
      enum: [redshift, snowflake, auto]
      default: auto

    ActivationStrategy:
      type: string
      enum: [auto, zero_etl, copy, iceberg_rest]
      description: |
        `auto` lets the adapter pick its preferred path
        (Redshift → zero_etl then copy; Snowflake → iceberg_rest).

    StorageMode:
      type: string
      enum: [materialized_copy, zero_copy]
      description: |
        Redshift activations are materialized copies (zero-ETL/COPY).
        Snowflake activations are zero-copy catalog registrations.

    ActivateRequest:
      type: object
      required: [gold_table, warehouse, gold_uri]
      additionalProperties: false
      properties:
        gold_table:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
          examples: [asset_daily_metrics]
        warehouse:
          $ref: "#/components/schemas/WarehouseId"
        gold_uri:
          type: string
          minLength: 1
          description: S3 or local file URI of the gold table directory
          examples:
            - s3://prism-gold/gold/asset_daily_metrics/
            - file:///data/lakehouse/gold/asset_daily_metrics
        strategy:
          $ref: "#/components/schemas/ActivationStrategy"
          default: auto
        set_primary:
          type: boolean
          default: true
          description: Mark this warehouse as the primary route for the table

    ActivateResponse:
      type: object
      required:
        - activation_id
        - gold_table
        - warehouse
        - strategy_used
        - storage_mode
        - status
        - gold_uri
        - row_count
        - activated_at
      additionalProperties: false
      properties:
        activation_id:
          type: string
          pattern: '^act_[0-9a-f]{12}$'
        gold_table:
          type: string
        warehouse:
          $ref: "#/components/schemas/WarehouseId"
        strategy_used:
          type: string
          enum: [zero_etl, copy, iceberg_rest]
        storage_mode:
          $ref: "#/components/schemas/StorageMode"
        status:
          type: string
          enum: [active]
        gold_uri:
          type: string
        row_count:
          type: integer
          minimum: 0
        activated_at:
          type: string
          format: date-time
        primary:
          type: boolean

    QueryRequest:
      type: object
      required: [table]
      additionalProperties: false
      properties:
        table:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
        warehouse:
          $ref: "#/components/schemas/WarehouseTarget"
        sql:
          type: string
          description: >
            Optional SELECT. Defaults to `SELECT * FROM <table> ORDER BY 1 LIMIT <limit>`.
            Only single SELECT statements are accepted.
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 100

    QueryResponse:
      type: object
      required:
        - table
        - warehouse
        - columns
        - rows
        - row_count
        - storage_mode
      additionalProperties: false
      properties:
        table:
          type: string
        warehouse:
          $ref: "#/components/schemas/WarehouseId"
        columns:
          type: array
          items:
            type: string
        rows:
          type: array
          items:
            type: array
            items: true
          description: Row-major values aligned to `columns`
        row_count:
          type: integer
          minimum: 0
        storage_mode:
          $ref: "#/components/schemas/StorageMode"
        sql:
          type: string

    RoutingEntry:
      type: object
      required: [warehouse, status, storage_mode, primary, gold_uri]
      additionalProperties: false
      properties:
        warehouse:
          $ref: "#/components/schemas/WarehouseId"
        status:
          type: string
          enum: [active]
        storage_mode:
          $ref: "#/components/schemas/StorageMode"
        primary:
          type: boolean
        gold_uri:
          type: string
        strategy_used:
          type: string
        row_count:
          type: integer
          minimum: 0
        activation_id:
          type: string

    RoutingResponse:
      type: object
      required: [table, warehouses]
      additionalProperties: false
      properties:
        table:
          type: string
        primary_warehouse:
          oneOf:
            - $ref: "#/components/schemas/WarehouseId"
            - type: "null"
        warehouses:
          type: array
          items:
            $ref: "#/components/schemas/RoutingEntry"

    WarehouseInfo:
      type: object
      required: [id, display_name, preferred_strategy, storage_mode]
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/WarehouseId"
        display_name:
          type: string
        preferred_strategy:
          type: string
        storage_mode:
          $ref: "#/components/schemas/StorageMode"
        endpoint:
          type: string

    WarehouseListResponse:
      type: object
      required: [warehouses]
      additionalProperties: false
      properties:
        warehouses:
          type: array
          items:
            $ref: "#/components/schemas/WarehouseInfo"

    HealthResponse:
      type: object
      required: [status, service]
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [ok]
        service:
          type: string
          enum: [activation-gateway]
        mode:
          type: string
          enum: [mock, live]
        warehouses:
          type: object
          additionalProperties:
            type: string

    ErrorResponse:
      type: object
      required: [detail]
      additionalProperties: false
      properties:
        detail:
          type: string